requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-import-export==4.3.9
django-map-widgets==0.5.1
django-modeltranslation==0.19.16
django-nested-admin==4.1.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-simple-history==3.10.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
et_xmlfile==2.0.0
file-validator==1.0.9
filetype==1.1.0
flower==2.0.1
google-api-core==2.25.0
google-auth==2.40.3
google-geo-type==0.3.13
google-maps-routing==0.6.16
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
humanize==4.4.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mypy_extensions==1.1.0
numpy==2.3.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.7
psycopg-binary==3.2.7
puremagic==1.14
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
python-magic==0.4.27
python-monkey-business==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tabulate==0.9.0
termcolor==1.1.0
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.2
xlwt==1.3.0
